Skip processing the manifest, handle it as a blob
The action or performing a remote tag does not need to process any
data from the manifest, and as such we can treat it as an opaque
blob.

This fixes the underlying issue in #1, where parsing the manifest
as JSON and `JSON.stringify`-ing it was causing a mismatch that
some Docker APIs did not handle.